Fix lint errors
